Added tap detection logic.
Added the support to detect taps on the sensor.
A sysfs attribute named 'tap_enable' is exported, just to enable
or disable tap event generation.
Writing '1' to it will enable tap detection (default). Writing '0'
to it disables reporting taps.
The tap detection logic is not changed at the harware level, but
only at the driver's software level.