Fixed run-time power-management issue
The event BTN_TOUCH was meddling with the power management.
With this event set, the open() function of the input framework
is being called immediately after the execution of probe(), thus
incrementing the usage count of the device, indirectly affecting
the power-management(as pm_runtime_get_sync() is called in open())
Thus as probe returned, the device was in a resume state!

The event is now changed to BTN_SELECT, which had solved the problem.
The change is also reflected in user application as well.

Now the power-management is working as desired.