Added support to configure samples per sec
The driver now exports a sysfs attribute named 'samples_ps' to
configure the samples per second updation of the X, Y and Z values.

The following samples_sec are supported by the device :
 120, 64, 32, 16, 8 , 4, 2 and 1.

Writing the corresponding numeber to the file, configures the device
as such. Reading from the file reads the current value of the samples
per second.

Also, the tap detection will be disabled, if samples per sec is other
than 120. The driver will automatically clear the flag and stop
generating events for the same. Trying to enable tap detection for
samples per sec other than 120, leads to an error and shown in the
kernel log.